requirements.txt
psycopg2-binary>=2.9.9
pyinstaller>=6.3.0
